Added ASCII table for --list
